Crate dumplingh

Issue/PR exporter for GitHub.

dumplingh as а library

Data flow

Options::parse()
|> list_{pull_requests,issues}()
|> save_{to_file,data}()

Example

let out_path = "pulls.json";
let repo = "nabijaczleweli/cargo-update".parse().unwrap();
let pulls = list_pull_requests(&repo).unwrap();
save_to_file(out_path, &pulls, "pull requests").unwrap();

dumplingh as аn executable

This is just a very short synopsis of the manpage, so consult that for more data.

OPTIONS

Option 	Description

<REPO_SLUG> 	Repository to export issues and PRs for in the form <username>/<repo>.

--issues <ISSUES_FILE> 	File to export issues to, or ./<slug>-issues.json by default.

--pulls <PULLS_FILE> 	File to export pull requests to, or ./<slug>-pulls.json by default.

--no-issues 	Don't export issues.

--no-pulls 	Don't export pull requests.

--force 	Override existing files.

Modules

ops
Main functions doing actual work.

util
Module containing various utility functions.

Structs

Options
Representation of the application's all configurable values.

Enums

Error
Enum representing all possible ways the application can fail.