Enum dumplingh::Error

pub enum Error {
    Io {
        desc: &'static str,
        op: &'static str,
        more: Option<&'static str>,
    },
    Parse {
        tp: &'static str,
        wher: &'static str,
        more: Option<&'static str>,
    },
}

Enum representing all possible ways the application can fail.

Variants
Io

An I/O error occured.

This includes higher-level I/O errors like FS ones.

Fields of Io

desc: &'static str

The file the I/O operation regards.

op: &'static str

The failed operation.

This should be lowercase and imperative ("create", "open").

more: Option<&'static str>

Additional data.

Parse

A parsing error occured.

Fields of Parse

tp: &'static str

What failed to parse.

Something like "URL", "datetime".

wher: &'static str

Where the thing that failed to parse would go, were it to parse properly.

more: Option<&'static str>

Additional data.

Methods

impl Error

fn print_error<W: Write>(&self, err_out: &mut W)

Write the error message to the specified output stream.

Examples

let mut out = Vec::new();
Error::Io {
    desc: "network",
    op: "write",
    more: Some("full buffer"),
}.print_error(&mut out);
assert_eq!(&out[..],
           &b"Writing network failed: full buffer.\n"[..]);

fn exit_value(&self) -> i32

Get the executable exit value from an Error instance.

Examples

assert_eq!(Error::Io {
    desc: "",
    op: "",
    more: None,
}.exit_value(), 1);